adjust data downloader and dataset
